104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,16 +10,16 @@ int main(void)
 {
 
 	int count = 2;
-	long int pre_number = 2;
-	long int pre_pre_number = 1;
-	long int printed_number;
+	double pre_number = 2;
+	double pre_pre_number = 1;
+	double printed_number;
 
 	printf("1, ");
 	printf("2, ");
 	while (count <= 98)
 	{
 		printed_number = pre_number + pre_pre_number;
-		printf("%ld", printed_number);
+		printf("%.0f", printed_number);
 		if (count < 98)
 		{
 			printf(", ");
